Set up the file browser functionality for music
Everything is in place to use the file browser, but the app crashes for
some reason when switching from the ExtremeBeatAnnihilation activity to
the FileExplore activity. The FileExplore activity is known to work, as
I have tested launching the app with the FileExplore activity first and
then switching to the ExtremeBeatAnnihilation activity. When you go to
switch back to the FileExplore activity then, it works instead of
crashing like it does in the normal case of launching the app with
ExtremeBeatAnnihilation activity first.